backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
pandas==2.1.3
polars==0.19.12
faiss-cpu==1.7.4
chromadb==0.4.15
python-dateutil==2.8.2
pytz==2023.3
numpy==1.26.2
